Refactored code to separate UI elements and logic
