+++ tasks.json
@@ -1,7 +1,1 @@
-[
-  {
-    "title": "task",
-    "description": "task",
-    "completed": false
-  }
-]
+[]
